requirements.txt
-f https://extras.wxpython.org/wxPython4/extras/linux/gtk3/ubuntu-18.04
wxPython>=4.1.0; sys_platform == 'linux'
wxPython>=4.1.0; sys_platform != 'linux'
setuptools>=52.0.0
numpy>=1.19.1
opencv-python>=4.5.3.56
mock>=4.0.3
sphinx-py3doc-enhanced-theme>=2.4.0
lxml>=4.6.3
attrs>=21.2.0
appdirs>=1.4.4
cattrs>=1.1.0
docstring-parser>=0.10.0
typing-extensions>=3.5.0; python_version<'3.8'
